Move SubItemTable into common folder
