+++ templates/vue/src/components/modals/NodeModal/forms/ContentForm/common/SubItemTable.vue
